rename title.py to return_title.py
This is the only function within the file, may as well rename it accordingly.
